refactor: constants 내부 폴더명 리네이밍
