+++ src/constants/validationRules/createActivityValidationRules.ts
